# src/API_SAC/swagger.openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "SAC - API Gestion Presences NFC",
    "description": "API REST de Saintonge Access Control. Swagger UI reutilise automatiquement la session Express courante; pas de bouton Authorize necessaire. Chaque operation affiche son role minimum dans son resume et sa description.",
    "version": "2.1.0"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
...
  "paths": {
...
    "/admin/sessions": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "[ADMIN] Creer une session de cours",
        "x-min-role": "admin",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CourseSessionInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session creee",
            "schema": {
              "$ref": "#/definitions/CourseSession"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/responses/Forbidden"
          },
          "500": {
            "$ref": "#/responses/InternalError"
          }
        },
        "description": "Role minimum: admin."
      }
    },
...
    "/business-logs": {
      "get": {
        "tags": [
          "Business Logs"
        ],
        "summary": "[STAFF+] Lister les logs metier",
        "description": "Role minimum: personnel ou admin.\n\nRetourne les logs metier stockes en BDD avec filtres par evenement, niveau, entite et utilisateur.",
        "x-min-role": "staff",
        "parameters": [
          {
            "in": "query",
            "name": "take",
            "type": "integer",
            "required": false,
            "description": "Nombre de logs a retourner, maximum 500."
          },
          {
            "in": "query",
            "name": "userId",
            "type": "integer",
            "required": false
          },
          {
            "in": "query",
            "name": "event",
            "type": "string",
            "required": false
          },
          {
            "in": "query",
            "name": "level",
            "type": "string",
            "required": false,
            "enum": [
              "ERROR",
              "WARNING",
              "INFO",
              "VERBOSE"
            ]
          },
          {
            "in": "query",
            "name": "entityType",
            "type": "string",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Logs metier",
            "schema": {
              "type": "object",
              "properties": {
                "filters": {
                  "type": "object"
                },
                "options": {
                  "type": "object"
                },
                "logs": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/BusinessLog"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/responses/Forbidden"
          },
          "500": {
            "$ref": "#/responses/InternalError"
          }
        }
      }
    },
    "/classes/staff": {
      "get": {
        "tags": [
          "Classes"
        ],
        "summary": "[STAFF+] Classes et eleves",
        "description": "Role minimum: personnel ou admin.\n\nRetourne les classes avec leurs eleves et les indicateurs utiles: compte O365 lie, deja connecte si o365Email existe, derniere session navigateur connue, dernier scan, mails, compteurs de presence.",
        "x-min-role": "staff",
        "responses": {
          "200": {
            "description": "Classes visibles par les personnels",
            "schema": {
              "$ref": "#/definitions/StaffClassesResponse"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/responses/Forbidden"
          },
          "500": {
            "$ref": "#/responses/InternalError"
          }
        }
      }
    },
    "/classes/me": {
      "get": {
        "tags": [
          "Classes"
        ],
        "summary": "[STUDENT+] Classe de l'eleve connecte",
        "description": "Role minimum: eleve ou plus.\n\nRetourne les informations de classe du compte eleve connecte.",
        "x-min-role": "student",
        "x-role-note": "La route est accessible a partir du role student, mais le code retourne 403 si l'utilisateur n'est pas un eleve.",
        "responses": {
          "200": {
            "description": "Classe de l'eleve",
            "schema": {
              "$ref": "#/definitions/ClassMeResponse"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/responses/Forbidden"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
    "/documentation": {
      "get": {
        "tags": [
          "Documentation"
        ],
        "summary": "[ADMIN] Interface Swagger UI",
        "x-min-role": "admin",
        "responses": {
          "200": {
            "description": "Swagger UI"
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/responses/Forbidden"
          }
        },
        "description": "Role minimum: admin."
      }
    },
...
  },
...
  "responses": {
...
    "Unauthorized": {
      "description": "Utilisateur non authentifie",
      "schema": {
        "$ref": "#/definitions/ErrorResponse"
      }
    },
    "Forbidden": {
      "description": "Role insuffisant ou action refusee",
      "schema": {
        "$ref": "#/definitions/ErrorResponse"
      }
    },
    "NotFound": {
      "description": "Ressource introuvable",
      "schema": {
        "$ref": "#/definitions/ErrorResponse"
      }
    },
    "InternalError": {
      "description": "Erreur serveur",
      "schema": {
        "$ref": "#/definitions/ErrorResponse"
      }
    }
  },
  "definitions": {
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Role": {
      "type": "string",
      "enum": [
        "student",
        "teacher",
        "staff",
        "admin"
      ]
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/Role"
        },
        "o365Email": {
          "type": "string"
        },
        "edEmail": {
          "type": "string"
        },
        "edId": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "classId": {
          "type": "integer"
        }
      }
    },
...
    "ClassMeResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "code": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "edId": {
          "type": "string"
        },
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ClassMeStudent"
          }
        }
      }
    },
    "ClassMeStudent": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "o365AvatarB64": {
          "type": "string"
        },
        "edPhotoB64": {
          "type": "string"
        }
      }
    },
    "StaffClassesResponse": {
      "type": "object",
      "properties": {
        "classes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StaffClass"
          }
        }
      }
    },
    "StaffClass": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "edId": {
          "type": "integer"
        },
        "code": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "studentsCount": {
          "type": "integer"
        },
        "linkedCount": {
          "type": "integer"
        },
        "loggedInCount": {
          "type": "integer"
        },
        "scannedCount": {
          "type": "integer"
        },
        "sessionsCount": {
          "type": "integer"
        },
        "students": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StaffClassStudent"
          }
        }
      }
    },
    "StaffClassStudent": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "edId": {
          "type": "string"
        },
        "o365Email": {
          "type": "string"
        },
        "edEmail": {
          "type": "string"
        },
        "o365AvatarB64": {
          "type": "string"
        },
        "edPhotoB64": {
          "type": "string"
        },
        "hasO365AccountLinked": {
          "type": "boolean"
        },
        "hasLoggedIn": {
          "type": "boolean"
        },
        "lastLoginAt": {
          "type": "string",
          "format": "date-time"
        },
        "lastNfcScanAt": {
          "type": "string",
          "format": "date-time"
        },
        "attendanceRecordsCount": {
          "type": "integer"
        },
        "nfcScansCount": {
          "type": "integer"
        },
        "lastAttendance": {
          "type": "object"
        }
      }
    },
...
    "CourseSession": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "label": {
          "type": "string"
        },
        "matiere": {
          "type": "string"
        },
        "codeMatiere": {
          "type": "string"
        },
        "color": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "classId": {
          "type": "integer"
        },
        "roomId": {
          "type": "integer"
        },
        "teacherId": {
          "type": "integer"
        },
        "startTime": {
          "type": "string",
          "format": "date-time"
        },
        "endTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "CourseSessionInput": {
      "type": "object",
      "properties": {
        "label": {
          "type": "string"
        },
        "matiere": {
          "type": "string"
        },
        "codeMatiere": {
          "type": "string"
        },
        "color": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "classId": {
          "type": "integer"
        },
        "roomId": {
          "type": "integer"
        },
        "teacherId": {
          "type": "integer"
        },
        "startTime": {
          "type": "string",
          "format": "date-time"
        },
        "endTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "BusinessLog": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "event": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "level": {
          "type": "string",
          "enum": [
            "ERROR",
            "WARNING",
            "INFO",
            "VERBOSE"
          ]
        },
        "userId": {
          "type": "integer"
        },
        "entityType": {
          "type": "string"
        },
        "entityId": {
          "type": "string"
        },
        "metadata": {
          "type": "object"
        },
        "ipAddress": {
          "type": "string"
        },
        "userAgent": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    }
  },
...
}